# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Trading mode - paper (simulasi) atau live (real trading)
 */
export type BotConfigMode = (typeof BotConfigMode)[keyof typeof BotConfigMode];

export const BotConfigMode = {
  paper: "paper",
  live: "live",
} as const;

export interface BotConfig {
  /** Asset ticker symbol */
  ticker: string;
  /** Lower price bound (USDC) */
  gridLow: number;
  /** Upper price bound (USDC) */
  gridHigh: number;
  /** Number of grid levels */
  gridCount: number;
  /** Order size per level in USDC */
  orderSizeUsdc: number;
  /** Initial paper trading balance (USDC) */
  initialBalance: number;
  /** Price polling interval in milliseconds */
  pollIntervalMs: number;
  /** Trading mode - paper (simulasi) atau live (real trading) */
  mode: BotConfigMode;
}

export interface GridLevelStatus {
  price: number;
  isOpen: boolean;
  buyPrice: number;
  quantity: number;
  realizedPnl: number;
  tradeCount: number;
}

export type BotStatusMode = (typeof BotStatusMode)[keyof typeof BotStatusMode];

export const BotStatusMode = {
  paper: "paper",
  live: "live",
} as const;

export interface BotStatus {
  running: boolean;
  ticker: string;
  currentPrice?: number | null;
  balance: number;
  initialBalance: number;
  openPositions: number;
  totalLevels: number;
  realizedPnl: number;
  unrealizedPnl: number;
  totalPnl: number;
  fillCount: number;
  fetchCount: number;
  uptimeSeconds: number;
  inRange: boolean;
  fundingRate: string;
  volume24h: string;
  gridLow: number;
  gridHigh: number;
  gridCount: number;
  orderSizeUsdc: number;
  pollIntervalMs: number;
  mode: BotStatusMode;
  liveBalanceUsdc?: number | null;
  liveUpnl?: number | null;
  levels: GridLevelStatus[];
}

export type TradeType = (typeof TradeType)[keyof typeof TradeType];

export const TradeType = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export interface Trade {
  id: number;
  type: TradeType;
  ticker: string;
  price: number;
  quantity: number;
  pnl?: number | null;
  timestamp: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetBotTradesParams = {
  limit?: number;
};

export type GetBotTrades200 = {
  trades: Trade[];
};