# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/config:
    get:
      operationId: getBotConfig
      tags: [bot]
      summary: Get current bot configuration
      responses:
        "200":
          description: Current config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
    patch:
      operationId: updateBotConfig
      tags: [bot]
      summary: Update bot configuration (bot must be stopped first)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfig"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
        "400":
          description: Invalid config or bot is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bot/trades:
    get:
      operationId: getBotTrades
      tags: [bot]
      summary: Get trade history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: object
                properties:
                  trades:
                    type: array
                    items:
                      $ref: "#/components/schemas/Trade"
                required:
                  - trades
...
components:
  schemas:
...
    BotConfig:
      type: object
      properties:
        ticker:
          type: string
          description: Asset ticker symbol
        gridLow:
          type: number
          description: Lower price bound (USDC)
        gridHigh:
          type: number
          description: Upper price bound (USDC)
        gridCount:
          type: integer
          description: Number of grid levels
        orderSizeUsdc:
          type: number
          description: Order size per level in USDC
        initialBalance:
          type: number
          description: Initial paper trading balance (USDC)
        pollIntervalMs:
          type: integer
          description: Price polling interval in milliseconds
        mode:
          type: string
          enum: [paper, live]
          description: Trading mode - paper (simulasi) atau live (real trading)
      required:
        - ticker
        - gridLow
        - gridHigh
        - gridCount
        - orderSizeUsdc
        - initialBalance
        - pollIntervalMs
        - mode
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [BUY, SELL]
        ticker:
          type: string
        price:
          type: number
        quantity:
          type: number
        pnl:
          type: number
          nullable: true
        timestamp:
          type: string
      required:
        - id
        - type
        - ticker
        - price
        - quantity
        - timestamp
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error