# lib/api-zod/src/generated/types/botConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotConfigMode } from "./botConfigMode";

export interface BotConfig {
  /** Asset ticker symbol */
  ticker: string;
  /** Lower price bound (USDC) */
  gridLow: number;
  /** Upper price bound (USDC) */
  gridHigh: number;
  /** Number of grid levels */
  gridCount: number;
  /** Order size per level in USDC */
  orderSizeUsdc: number;
  /** Initial paper trading balance (USDC) */
  initialBalance: number;
  /** Price polling interval in milliseconds */
  pollIntervalMs: number;
  /** Trading mode - paper (simulasi) atau live (real trading) */
  mode: BotConfigMode;
}